requirements.txt
PyQt5
numpy
pillow
opencv-python